fix exmples to work with unittest
