+++ examples/README.md
@@ -1,0 +1,15 @@
+# Examples
+
+The examples can be run two different ways
+
+### Directly
+
+Most examples are simple Python programs that define a grammar and parse some strings. They will throw an exception and error-out if parsing failed. You can run them directly. 
+
+    python examples/antler_grammar.py
+
+> Be sure you are in the `pyparsing` project directory, not the `example` directory
+
+### Test Suite
+
+The test suite can run the examples for you
