+++ examples/SimpleCalc.py
@@ -32,7 +32,7 @@ debug_flag = False
 
 variables = {}
 
-from fourFn import BNF, exprStack, evaluate_stack
+from examples.fourFn import BNF, exprStack, evaluate_stack
 
 # from fourFn import BNF, exprStack, fn, opn
 # def evaluateStack( s ):
@@ -62,6 +62,7 @@ assignment = ident("varname") + "=" + arithExpr
 pattern = assignment | arithExpr
 
 if __name__ == "__main__":
+
     # input_string
     input_string = ""
