+++ examples/antlr_grammar.py
@@ -338,51 +338,3 @@ def antlrConverter(antlrGrammarTree):
         pyparsingRules[antlrRuleName] <<= pyparsingRule
 
     return pyparsingRules
-
-
-if __name__ == "__main__":
-
-    text = """\
-grammar SimpleCalc;
-
-options {
-    language = Python;
-}
-
-tokens {
-    PLUS     = '+' ;
-    MINUS    = '-' ;
-    MULT    = '*' ;
-    DIV    = '/' ;
-}
-
-/*------------------------------------------------------------------
- * PARSER RULES
- *------------------------------------------------------------------*/
-
-expr    : term ( ( PLUS | MINUS )  term )* ;
-
-term    : factor ( ( MULT | DIV ) factor )* ;
-
-factor    : NUMBER ;
-
-
-/*------------------------------------------------------------------
- * LEXER RULES
- *------------------------------------------------------------------*/
-
-NUMBER    : (DIGIT)+ ;
-
-/* WHITESPACE : ( '\t' | ' ' | '\r' | '\n'| '\u000C' )+     { $channel = HIDDEN; } ; */
-
-fragment DIGIT    : '0'..'9' ;
-
-"""
-
-    grammar().validate()
-    antlrGrammarTree = grammar().parseString(text)
-    print(antlrGrammarTree.dump())
-    pyparsingRules = antlrConverter(antlrGrammarTree)
-    pyparsingRule = pyparsingRules["expr"]
-    pyparsingTree = pyparsingRule.parseString("2 - 5 * 42 + 7 / 25")
-    print(pyparsingTree.dump())
