+++ examples/invRegex.py
@@ -305,5 +305,4 @@ def main():
         print("")
 
 
-if __name__ == "__main__":
-    main()
+main()
