+++ examples/jsonParser.py
@@ -63,50 +63,3 @@ jsonObject << pp.Dict(LBRACE + pp.Optional(jsonMembers) + RBRACE)
 
 jsonComment = pp.cppStyleComment
 jsonObject.ignore(jsonComment)
-
-
-if __name__ == "__main__":
-    testdata = """
-    {
-        "glossary": {
-            "title": "example glossary",
-            "GlossDiv": {
-                "title": "S",
-                "GlossList":
-                    {
-                    "ID": "SGML",
-                    "SortAs": "SGML",
-                    "GlossTerm": "Standard Generalized Markup Language",
-                    "TrueValue": true,
-                    "FalseValue": false,
-                    "Gravity": -9.8,
-                    "LargestPrimeLessThan100": 97,
-                    "AvogadroNumber": 6.02E23,
-                    "EvenPrimesGreaterThan2": null,
-                    "PrimesLessThan10" : [2,3,5,7],
-                    "Acronym": "SGML",
-                    "Abbrev": "ISO 8879:1986",
-                    "GlossDef": "A meta-markup language, used to create markup languages such as DocBook.",
-                    "GlossSeeAlso": ["GML", "XML", "markup"],
-                    "EmptyDict" : {},
-                    "EmptyList" : []
-                    }
-            }
-        }
-    }
-    """
-
-    results = jsonObject.parseString(testdata)
-    results.pprint()
-    print()
-
-    def testPrint(x):
-        print(type(x), repr(x))
-
-    print(list(results.glossary.GlossDiv.GlossList.keys()))
-    testPrint(results.glossary.title)
-    testPrint(results.glossary.GlossDiv.GlossList.ID)
-    testPrint(results.glossary.GlossDiv.GlossList.FalseValue)
-    testPrint(results.glossary.GlossDiv.GlossList.Acronym)
-    testPrint(results.glossary.GlossDiv.GlossList.EvenPrimesGreaterThan2)
-    testPrint(results.glossary.GlossDiv.GlossList.PrimesLessThan10)
