+++ examples/statemachine/video_demo.py
@@ -15,38 +15,45 @@ class Video(videostate.VideoStateMixin):
 
 
 # ==== main loop - a REPL ====
+if __name__ == "__main__":
 
-v = Video("Die Hard.mp4")
+    v = Video("Die Hard.mp4")
 
-while True:
-    print(v.state)
-    cmd = (
-        input("Command ({})> ".format("/".join(videostate.VideoState.transition_names)))
-        .lower()
-        .strip()
-    )
-    if not cmd:
-        continue
+    while True:
+        print(v.state)
+        cmd = (
+            input(
+                "Command ({})> ".format(
+                    "/".join(videostate.VideoState.transition_names)
+                )
+            )
+            .lower()
+            .strip()
+        )
+        if not cmd:
+            continue
 
-    if cmd in ("?", "h", "help"):
-        print("enter a transition {!r}".format(videostate.VideoState.transition_names))
-        print(" q - quit")
-        print(" ?, h, help - this message")
-        continue
+        if cmd in ("?", "h", "help"):
+            print(
+                "enter a transition {!r}".format(videostate.VideoState.transition_names)
+            )
+            print(" q - quit")
+            print(" ?, h, help - this message")
+            continue
 
-    # quitting out
-    if cmd.startswith("q"):
-        break
+        # quitting out
+        if cmd.startswith("q"):
+            break
 
-    # get transition function for given command
-    state_transition_fn = getattr(v, cmd, None)
+        # get transition function for given command
+        state_transition_fn = getattr(v, cmd, None)
 
-    if state_transition_fn is None:
-        print("???")
-        continue
+        if state_transition_fn is None:
+            print("???")
+            continue
 
-    # invoke the input transition, handle invalid commands
-    try:
-        state_transition_fn()
-    except videostate.VideoState.InvalidTransitionException as e:
-        print(e)
+        # invoke the input transition, handle invalid commands
+        try:
+            state_transition_fn()
+        except videostate.VideoState.InvalidTransitionException as e:
+            print(e)
