+++ examples/verilogParse.py
@@ -888,108 +888,80 @@ def Verilog_BNF():
     return verilogbnf
 
 
-def test(strng):
-    tokens = []
-    try:
-        tokens = Verilog_BNF().parseString(strng)
-    except ParseException as err:
-        print(err.line)
-        print(" " * (err.column - 1) + "^")
-        print(err)
-    return tokens
+if __name__ == "__main__":
 
+    print("Verilog parser test (V %s)" % __version__)
+    print(" - using pyparsing version", pyparsing.__version__)
+    print(" - using Python version", sys.version)
+    if packratOn:
+        print(" - using packrat parsing")
+    print()
 
-# ~ if __name__ == "__main__":
-if 0:
-    import pprint
+    import os
+    import gc
 
-    toptest = """
-        module TOP( in, out );
-        input [7:0] in;
-        output [5:0] out;
-        COUNT_BITS8 count_bits( .IN( in ), .C( out ) );
-        endmodule"""
-    pprint.pprint(test(toptest).asList())
+    failCount = 0
+    Verilog_BNF()
+    numlines = 0
+    startTime = time.clock()
+    fileDir = "verilog"
+    # ~ fileDir = "verilog/new"
+    # ~ fileDir = "verilog/new2"
+    # ~ fileDir = "verilog/new3"
+    allFiles = [f for f in os.listdir(fileDir) if f.endswith(".v")]
+    # ~ allFiles = [ "list_path_delays_test.v" ]
+    # ~ allFiles = [ "escapedIdent.v" ]
+    # ~ allFiles = filter( lambda f : f.startswith("a") and f.endswith(".v"), os.listdir(fileDir) )
+    # ~ allFiles = filter( lambda f : f.startswith("c") and f.endswith(".v"), os.listdir(fileDir) )
+    # ~ allFiles = [ "ff.v" ]
 
-else:
+    pp = pprint.PrettyPrinter(indent=2)
+    totalTime = 0
+    for vfile in allFiles:
+        gc.collect()
+        fnam = fileDir + "/" + vfile
+        infile = open(fnam)
+        filelines = infile.readlines()
+        infile.close()
+        print(fnam, len(filelines), end=" ")
+        numlines += len(filelines)
+        teststr = "".join(filelines)
+        time1 = time.clock()
+        tokens = test(teststr)
+        time2 = time.clock()
+        elapsed = time2 - time1
+        totalTime += elapsed
+        if len(tokens):
+            print("OK", elapsed)
+            # ~ print "tokens="
+            # ~ pp.pprint( tokens.asList() )
+            # ~ print
 
-    def main():
-        print("Verilog parser test (V %s)" % __version__)
-        print(" - using pyparsing version", pyparsing.__version__)
-        print(" - using Python version", sys.version)
-        if packratOn:
-            print(" - using packrat parsing")
-        print()
-
-        import os
-        import gc
-
-        failCount = 0
-        Verilog_BNF()
-        numlines = 0
-        startTime = time.clock()
-        fileDir = "verilog"
-        # ~ fileDir = "verilog/new"
-        # ~ fileDir = "verilog/new2"
-        # ~ fileDir = "verilog/new3"
-        allFiles = [f for f in os.listdir(fileDir) if f.endswith(".v")]
-        # ~ allFiles = [ "list_path_delays_test.v" ]
-        # ~ allFiles = [ "escapedIdent.v" ]
-        # ~ allFiles = filter( lambda f : f.startswith("a") and f.endswith(".v"), os.listdir(fileDir) )
-        # ~ allFiles = filter( lambda f : f.startswith("c") and f.endswith(".v"), os.listdir(fileDir) )
-        # ~ allFiles = [ "ff.v" ]
-
-        pp = pprint.PrettyPrinter(indent=2)
-        totalTime = 0
-        for vfile in allFiles:
-            gc.collect()
-            fnam = fileDir + "/" + vfile
-            infile = open(fnam)
-            filelines = infile.readlines()
-            infile.close()
-            print(fnam, len(filelines), end=" ")
-            numlines += len(filelines)
-            teststr = "".join(filelines)
-            time1 = time.clock()
-            tokens = test(teststr)
-            time2 = time.clock()
-            elapsed = time2 - time1
-            totalTime += elapsed
-            if len(tokens):
-                print("OK", elapsed)
-                # ~ print "tokens="
-                # ~ pp.pprint( tokens.asList() )
-                # ~ print
-
-                ofnam = fileDir + "/parseOutput/" + vfile + ".parsed.txt"
-                outfile = open(ofnam, "w")
-                outfile.write(teststr)
-                outfile.write("\n")
-                outfile.write("\n")
-                outfile.write(pp.pformat(tokens.asList()))
-                outfile.write("\n")
-                outfile.close()
-            else:
-                print("failed", elapsed)
-                failCount += 1
-                for i, line in enumerate(filelines, 1):
-                    print("%4d: %s" % (i, line.rstrip()))
-        endTime = time.clock()
-        print("Total parse time:", totalTime)
-        print("Total source lines:", numlines)
-        print("Average lines/sec:", ("%.1f" % (float(numlines) / (totalTime + 0.05))))
-        if failCount:
-            print("FAIL - %d files failed to parse" % failCount)
+            ofnam = fileDir + "/parseOutput/" + vfile + ".parsed.txt"
+            outfile = open(ofnam, "w")
+            outfile.write(teststr)
+            outfile.write("\n")
+            outfile.write("\n")
+            outfile.write(pp.pformat(tokens.asList()))
+            outfile.write("\n")
+            outfile.close()
         else:
-            print("SUCCESS - all files parsed")
-
-        return 0
+            print("failed", elapsed)
+            failCount += 1
+            for i, line in enumerate(filelines, 1):
+                print("%4d: %s" % (i, line.rstrip()))
+    endTime = time.clock()
+    print("Total parse time:", totalTime)
+    print("Total source lines:", numlines)
+    print("Average lines/sec:", ("%.1f" % (float(numlines) / (totalTime + 0.05))))
+    if failCount:
+        print("FAIL - %d files failed to parse" % failCount)
+    else:
+        print("SUCCESS - all files parsed")
 
     # ~ from line_profiler import LineProfiler
     # ~ from pyparsing import ParseResults
     # ~ lp = LineProfiler(ParseResults.__init__)
-
-    main()
 
     # ~ lp.print_stats()
     # ~ import hotshot
